Change price type from Double to BigDecimal
Updated the `price` field type to `BigDecimal` in `ProductUpdateRequest`, `ProductCreateRequest`, and `ProductDTO` for better precision. Adjusted import statements accordingly.

+++ src/main/java/tech/nuqta/handihub/product/dto/ProductDTO.java
@@ -1,6 +1,7 @@
 package tech.nuqta.handihub.product.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -24,7 +25,7 @@ public record ProductDTO(
         boolean isHandmade,
         boolean madeToOrder,
         List<String> images,
-        Double price,
+        BigDecimal price,
         Integer quantity,
         Long userId,
         Long categoryId
